Send GameServer's Address to all PigServers, Implement Stone Falling at GameServer
Remove portNo from GameServer's command-line arguments,
since it can be read from the configuration.

When a pig falls on a stone, it informs GameServer.
GameServer then selects a random cell for the stone to fall on;
if it is a pig, kills the pig; if it is a stone, calls the same
method for the cell of that stone.