Add nonNull filter to neighbor streams, Remove NeighborCellUpdater
Since the neigbors list contains null objects, it would cause an NPE
whenever streaming data of the neighbors list. Thus, added a nonNull
filter to fix that.

NeighborCellUpdater.java was taking time to update the originator pig.
Due to that, originator would not get the updated location of the
neighbor pig, and mess things up.
So, now, takeShelter() now returns a boolean so as to update serially.